feat: パラメータ取得ロジック実装
